Confidence interval associated with discovery GWAS conditional likelihood methods

cl_interval is a function that allows the user to obtain a confidence interval for the adjusted association estimates of significant SNPs, which have been obtained through the implementation of conditional_likelihood. This function produces one confidence interval for each significant SNP, based on the approach suggested in Ghosh et al. (2008).

cl_interval(summary_data, alpha, conf_level = 0.95)

Arguments

summary_data
A data frame containing summary statistics from the discovery GWAS. It must have three columns with column names rsid, beta and se, respectively, and all columns must contain numerical values.

alpha
A numerical value which specifies the desired genome-wide significance threshold.

conf_level
A numerical value between 0 and 1 which determines the confidence interval to be computed. The default setting is 0.95 which results in the calculation of a 95% confidence interval for the adjusted association estimate for each SNP.

Value

A data frame which combines the output of conditional_likelihood with two additional columns, namely lower and upper, containing the lower and upper bounds of the required confidence interval for each significant SNP, respectively.

References

Ghosh, A., Zou, F., & Wright, F. A. (2008). Estimating odds ratios in genome scans: an approximate conditional likelihood approach. American journal of human genetics, 82(5), 1064\(-\)1074. https://doi.org/10.1016/j.ajhg.2008.03.002

See also