Standard errors of adjusted discovery GWAS estimates via parametric bootstrap

se_adjust is a function which allows the user to obtain approximate standard errors of adjusted association estimates, by means of parametric bootstrapping. Standard errors can be evaluated for estimates which have been corrected with the Empirical Bayes method, FDR Inverse Quantile Transformation method or the bootstrap method. Note that in comparison to the other functions in this package, this function can be computationally intensive and take a several minutes to run, depending on the size of the data set, the method and the number of bootstraps chosen.

se_adjust(summary_data, method, n_boot = 100)

Arguments

summary_data

A data frame containing summary statistics from the discovery GWAS. It must have three columns with column names rsid, beta and se, respectively, and columns beta and se must contain numerical values. Each row must correspond to a unique SNP, identified by rsid.

method

A string specifying the function to be implemented on each of the bootstrap samples. It should take the form "BR_ss", "empirical_bayes" or "FDR_IQT".

n_boot

A numerical value which determines the number of bootstrap repetitions to be used. it must be greater than 1. The default value is 100.

Value

A data frame which combines the output of the chosen method with an additional column, namely adj_se. This column provides the standard errors of the adjusted association estimates for each SNP.

See also