FDR IQT method for use with discovery GWAS

FDR_IQT is a function which uses summary statistics to reduce Winner's Curse bias in SNP-trait association estimates, obtained from a discovery GWAS. The function implements the FDR Inverse Quantile Transformation method described in Bigdeli et al. (2016), which was established for this purpose.

FDR_IQT(summary_data, min_pval = 1e-15)

Arguments

summary_data

A data frame containing summary statistics from the discovery GWAS. It must have three columns with column names rsid, beta and se, respectively, and columns beta and se must contain numerical values. Each row must correspond to a unique SNP, identified by the numerical value rsid.

min_pval

A numerical value whose purpose is to avoid zero \(p\)-values as this introduces issues when qnorm() is applied. Any SNP for which its computed \(p\)-value is found to be less than min_pval is merely re-assigned min_pval as its \(p\)-value and the method proceeds. By definition, the method makes no adjustment to the association estimate of a SNP for which this has occurred with the presumption that in general, estimates of SNPs with \(z > 8\) are not biased. The default value is min_pval = 1e-15.

Value

A data frame with the inputted summary data occupying the first three columns. The new adjusted association estimates for each SNP are returned in the fourth column, namely beta_FIQT. The SNPs are contained in this data frame according to their significance, with the most significant SNP, i.e. the SNP with the largest absolute \(z\)-statistic, now located in the first row of the data frame.

References

Bigdeli, T. B., Lee, D., Webb, B. T., Riley, B. P., Vladimirov, V. I., Fanous, A. H., Kendler, K. S., & Bacanu, S. A. (2016). A simple yet accurate correction for winner's curse can predict signals discovered in much larger genome scans. Bioinformatics (Oxford, England), 32(17), 2598\(-\)2603. https://doi.org/10.1093/bioinformatics/btw303

See also